backend/requirements.txt
fastapi==0.136.0
librosa==0.11.0
numpy==2.4.4
openai_whisper==20250625
pydantic==2.13.3
rapidfuzz==3.14.5
scipy==1.17.1
